+++ ontimize-boot-keycloak/src/main/java/com/ontimize/boot/keycloak/OntimizeKeycloakConfiguration.java
@@ -1,27 +1,25 @@
 package com.ontimize.boot.keycloak;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import com.ontimize.jee.server.security.keycloak.IOntimizeKeycloakConfiguration;
 
+@ConfigurationProperties(prefix = "ontimize.security.keycloak")
 public class OntimizeKeycloakConfiguration implements IOntimizeKeycloakConfiguration {
-	@Value("${ontimize.security.ignore-paths:}")
-	private String[] ignorePaths;
-
 	@Value("${ontimize.security.keycloak.public-client:false}")
 	private Boolean publicClient;
 
 	@Value("${ontimize.security.keycloak.use-resource-role-mappings:false}")
 	private Boolean useResourceRoleMappings;
 
-	@Override
-	public String[] getIgnorePaths() {
-		return ignorePaths;
-	}
+	@Value("${ontimize.security.keycloak.realms-provider:}")
+	private String realmsProvider;
 
-	public void setIgnorePaths(String[] ignorePaths) {
-		this.ignorePaths = ignorePaths;
-	}
+	private List<OntimizeKeycloakRoleConfig> roles = new ArrayList<>();
 
 	@Override
 	public Boolean isPublicClient() {
@@ -39,5 +37,21 @@ public class OntimizeKeycloakConfiguration implements IOntimizeKeycloakConfigura
 
 	public void setUseResourceRoleMappings(Boolean useResourceRoleMappings) {
 		this.useResourceRoleMappings = useResourceRoleMappings;
+	}
+
+	public String getRealmsProvider() {
+		return realmsProvider;
+	}
+
+	public void setRealmsProvider(String realmsProvider) {
+		this.realmsProvider = realmsProvider;
+	}
+
+	public List<OntimizeKeycloakRoleConfig> getRoles() {
+		return roles;
+	}
+
+	public void setRoles(List<OntimizeKeycloakRoleConfig> roles) {
+		this.roles = roles;
 	}
 }