+++ ontimize-boot-keycloak/src/main/java/com/ontimize/boot/keycloak/OntimizeKeycloakWebSecurityConfigurerAdapter.java
@@ -2,8 +2,10 @@ package com.ontimize.boot.keycloak;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.keycloak.adapters.springsecurity.KeycloakConfiguration;
 import org.keycloak.adapters.springsecurity.KeycloakSecurityComponents;
@@ -11,11 +13,14 @@ import org.keycloak.adapters.springsecurity.authentication.KeycloakAuthenticatio
 import org.keycloak.adapters.springsecurity.config.KeycloakWebSecurityConfigurerAdapter;
 import org.keycloak.adapters.springsecurity.filter.KeycloakAuthenticationProcessingFilter;
 import org.keycloak.adapters.springsecurity.filter.KeycloakPreAuthActionsFilter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
@@ -41,17 +46,22 @@ import org.springframework.security.web.authentication.session.SessionAuthentica
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
+import com.ontimize.jee.common.security.XMLClientUtilities;
 import com.ontimize.jee.server.dao.IOntimizeDaoSupport;
 import com.ontimize.jee.server.requestfilter.OntimizePathMatcher;
 import com.ontimize.jee.server.security.DatabaseRoleInformationService;
 import com.ontimize.jee.server.security.ISecurityRoleInformationService;
 import com.ontimize.jee.server.security.SecurityConfiguration;
 import com.ontimize.jee.server.security.authorization.DefaultOntimizeAuthorizator;
+import com.ontimize.jee.server.security.authorization.DefaultRoleProvider;
+import com.ontimize.jee.server.security.authorization.IRoleProvider;
 import com.ontimize.jee.server.security.authorization.ISecurityAuthorizator;
 import com.ontimize.jee.server.security.authorization.OntimizeAccessDecisionVoter;
+import com.ontimize.jee.server.security.authorization.Role;
 import com.ontimize.jee.server.security.keycloak.IOntimizeKeycloakConfiguration;
 import com.ontimize.jee.server.security.keycloak.OntimizeKeycloakConfigResolver;
 import com.ontimize.jee.server.security.keycloak.OntimizeKeycloakUserDetailsAuthenticationProvider;
+import com.ontimize.jee.server.security.keycloak.OntimizeKeycloakRoleProvider;
 import com.ontimize.jee.server.security.keycloak.admin.IUserManagement;
 import com.ontimize.jee.server.security.keycloak.admin.UserManagementKeycloakImpl;
 
@@ -61,8 +71,13 @@ import com.ontimize.jee.server.security.keycloak.admin.UserManagementKeycloakImp
 @ComponentScan(basePackageClasses = KeycloakSecurityComponents.class,
 		excludeFilters = @ComponentScan.Filter(type = FilterType.REGEX, pattern = "org.keycloak.adapters.springsecurity.management.HttpSessionManager"))
 public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSecurityConfigurerAdapter {
+	private static final Logger logger = LoggerFactory.getLogger(OntimizeKeycloakWebSecurityConfigurerAdapter.class);
+	
 	@Value("${ontimize.security.ignore-paths:}")
 	private String[] ignorePaths;
+	
+	@Autowired
+	private OntimizeKeycloakConfiguration keycloakConfiguration;
 
 	@Autowired
 	@Qualifier("pathMatcherIgnorePaths")
@@ -70,7 +85,27 @@ public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSec
 
 	@Override
 	protected AuthenticationEntryPoint authenticationEntryPoint() throws Exception {
-		return new OntimizeKeycloakAuthenticationEntryPoint(adapterDeploymentContext(), this.pathMatcherIgnorePaths);
+		final OntimizeKeycloakTenantValidator tenantValidator = new OntimizeKeycloakTenantValidator(this.pathMatcherIgnorePaths,
+				"list".equals(this.keycloakConfiguration.getRealmsProvider()) || "custom".equals(this.keycloakConfiguration.getRealmsProvider()));
+		return new OntimizeKeycloakAuthenticationEntryPoint(adapterDeploymentContext(), tenantValidator);
+	}
+
+	@Bean
+	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "custom", matchIfMissing = false)
+	public IOntimizeKeycloakConfiguration createOntimizeKeycloakConfiguration() {
+		return new OntimizeKeycloakConfiguration();
+	}
+
+	@Bean
+	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "list", matchIfMissing = false)
+	public IOntimizeKeycloakConfiguration createOntimizeKeycloakMultiTenantConfiguration() {
+		return new OntimizeKeycloakMultiTenantConfiguration();
+	}
+
+	@Bean
+	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "default", matchIfMissing = true)
+	public IOntimizeKeycloakConfiguration createOntimizeKeycloakSingleTenantConfiguration() {
+		return new OntimizeKeycloakSingleTenantConfiguration();
 	}
 
 	@Bean
@@ -111,20 +146,20 @@ public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSec
 
 	@Bean
 	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "custom", matchIfMissing = false)
-	public IOntimizeKeycloakConfiguration createOntimizeKeycloakConfiguration() {
-		return new OntimizeKeycloakConfiguration();
+	public OntimizeKeycloakTenantValidator createOntimizeKeycloakTenantValidator() {
+		return new OntimizeKeycloakTenantValidator(this.pathMatcherIgnorePaths, true);
 	}
 
 	@Bean
 	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "list", matchIfMissing = false)
-	public IOntimizeKeycloakConfiguration createOntimizeKeycloakMultiTenantConfiguration() {
-		return new OntimizeKeycloakMultiTenantConfiguration();
+	public OntimizeKeycloakTenantValidator createOntimizeKeycloakMultiTenantValidator() {
+		return new OntimizeKeycloakTenantValidator(this.pathMatcherIgnorePaths, true);
 	}
 
 	@Bean
 	@ConditionalOnProperty(name = "ontimize.security.keycloak.realms-provider", havingValue = "default", matchIfMissing = true)
-	public IOntimizeKeycloakConfiguration createOntimizeKeycloakSingleTenantConfiguration() {
-		return new OntimizeKeycloakSingleTenantConfiguration();
+	public OntimizeKeycloakTenantValidator createOntimizeKeycloakSingleTenantValidator() {
+		return new OntimizeKeycloakTenantValidator(this.pathMatcherIgnorePaths, false);
 	}
 
 	@Bean
@@ -190,27 +225,50 @@ public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSec
 	}
 
 	@Bean
-	public ISecurityRoleInformationService roleInformationService() {
-		RoleInformationServiceConfig config = this.roleInformationServiceConfig();
-		DatabaseRoleInformationService roleInformationService = new DatabaseRoleInformationService();
+	@ConditionalOnProperty(name = "ontimize.security.role-information-service.role-repository")
+	public ISecurityRoleInformationService roleInformationService(final RoleInformationServiceConfig config) {
+		final DatabaseRoleInformationService roleInformationService = new DatabaseRoleInformationService();
 
-		if (config.getRoleRepository() != null) {
-			Object roleDao = this.getApplicationContext().getBean(config.getRoleRepository());
-			if (roleDao instanceof IOntimizeDaoSupport) {
-				roleInformationService.setProfileRepository((IOntimizeDaoSupport) roleDao);
-			}
-
-			roleInformationService.setRoleNameColumn(config.getRoleNameColumn());
-			roleInformationService.setServerPermissionQueryId(config.getServerPermissionQueryId());
-			roleInformationService.setServerPermissionKeyColumn(config.getServerPermissionNameColumn());
-			roleInformationService.setClientPermissionQueryId(config.getClientPermissionQueryId());
-			roleInformationService.setClientPermissionColumn(config.getClientPermissionColumn());
+		final Object roleDao = this.getApplicationContext().getBean(config.getRoleRepository());
+		if (roleDao instanceof IOntimizeDaoSupport) {
+			roleInformationService.setProfileRepository((IOntimizeDaoSupport) roleDao);
 		}
+
+		roleInformationService.setRoleNameColumn(config.getRoleNameColumn());
+		roleInformationService.setServerPermissionQueryId(config.getServerPermissionQueryId());
+		roleInformationService.setServerPermissionKeyColumn(config.getServerPermissionNameColumn());
+		roleInformationService.setClientPermissionQueryId(config.getClientPermissionQueryId());
+		roleInformationService.setClientPermissionColumn(config.getClientPermissionColumn());
+
 		return roleInformationService;
 	}
 
+	@Bean
+	public IRoleProvider roleProvider(final ApplicationContext context, final RoleInformationServiceConfig config) {
+		if (config.getRoleRepository() != null) {
+			return new DefaultRoleProvider(context);
+		} else {
+			final OntimizeKeycloakRoleProvider roleProvider = new OntimizeKeycloakRoleProvider();
+
+			for (final OntimizeKeycloakRoleConfig roleConfig : this.keycloakConfiguration.getRoles()) {
+				Map<String, String> clientPermissions = new HashMap<>();
+				if (roleConfig.getClientPermissions() != null) {
+					try {
+						clientPermissions = XMLClientUtilities.buildClientPermissions(new StringBuffer(roleConfig.getClientPermissions()));
+					} catch (Exception e) {
+						OntimizeKeycloakWebSecurityConfigurerAdapter.logger.error("Error loading client permissions for role {}", roleConfig.getName(), e);
+					}
+				}
+
+				roleProvider.addRole(new Role(roleConfig.getName(), roleConfig.getServerPermissions(), clientPermissions));
+			}
+
+			return roleProvider;
+		}
+	}
+
 	public FilterSecurityInterceptor filterInvocationInterceptor() throws Exception {
-		FilterSecurityInterceptor filterInvocationInterceptor = new FilterSecurityInterceptor();
+		final FilterSecurityInterceptor filterInvocationInterceptor = new FilterSecurityInterceptor();
 		filterInvocationInterceptor.setObserveOncePerRequest(true);
 		filterInvocationInterceptor.setAuthenticationManager(this.authenticationManager());
 		filterInvocationInterceptor.setAccessDecisionManager(this.accessDecisionManager());
@@ -220,7 +278,7 @@ public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSec
 
 	@Bean
 	public FilterInvocationSecurityMetadataSource filterInvocationSecurityMetadataSource() {
-		LinkedHashMap<RequestMatcher, Collection<ConfigAttribute>> requestMap = new LinkedHashMap<>();
+		final LinkedHashMap<RequestMatcher, Collection<ConfigAttribute>> requestMap = new LinkedHashMap<>();
 		requestMap.put(new AntPathRequestMatcher("/**/*"), SecurityConfig.createList("NONE_ENTER_WITHOUT_AUTH"));
 
 		return new ExpressionBasedFilterInvocationSecurityMetadataSource(requestMap,
@@ -228,9 +286,11 @@ public class OntimizeKeycloakWebSecurityConfigurerAdapter extends KeycloakWebSec
 	}
 
 	@Bean
-	public SecurityConfiguration securityConfiguration() {
-		SecurityConfiguration securityConfiguration = new SecurityConfiguration();
-		securityConfiguration.setRoleInformationService(this.roleInformationService());
+	public SecurityConfiguration securityConfiguration(final RoleInformationServiceConfig config) {
+		final SecurityConfiguration securityConfiguration = new SecurityConfiguration();
+		if (config.getRoleRepository() != null) {
+			securityConfiguration.setRoleInformationService(this.roleInformationService(config));
+		}
 		securityConfiguration.setAuthorizator(this.ontimizeAuthorizator());
 		return securityConfiguration;
 	}